LayoutTests/platform/glib/imported/w3c/web-platform-tests/fetch/cross-origin-resource-policy/fetch.https.any-expected.txt

CONSOLE MESSAGE: Cancelled load to https://not-web-platform.test:9443/fetch/cross-origin-resource-policy/resources/hello.py?corp=same-origin because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://not-web-platform.test:9443/fetch/cross-origin-resource-policy/resources/hello.py?corp=same-origin due to access control checks.
CONSOLE MESSAGE: Cancelled load to https://not-web-platform.test:9443/fetch/cross-origin-resource-policy/resources/hello.py?corp=same-site because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://not-web-platform.test:9443/fetch/cross-origin-resource-policy/resources/hello.py?corp=same-site due to access control checks.
CONSOLE MESSAGE: Cancelled load to https://not-web-platform.test:9443/fetch/cross-origin-resource-policy/resources/hello.py?corp=same-origin because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://not-web-platform.test:9443/fetch/cross-origin-resource-policy/resources/hello.py?corp=same-origin due to access control checks.
CONSOLE MESSAGE: Cross-origin redirection to https://web-platform.test:9443/fetch/cross-origin-resource-policy/resources/hello.py?corp=same-origin denied by Cross-Origin Resource Sharing policy: Cancelled load to https://not-web-platform.test:9443/fetch/cross-origin-resource-policy/resources/redirect.py?corp=same-origin&redirectTo=https%3A%2F%2Fweb-platform.test%3A9443%2Ffetch%2Fcross-origin-resource-policy%2Fresources%2Fhello.py%3Fcorp%3Dsame-origin because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://not-web-platform.test:9443/fetch/cross-origin-resource-policy/resources/redirect.py?corp=same-origin&redirectTo=https%3A%2F%2Fweb-platform.test%3A9443%2Ffetch%2Fcross-origin-resource-policy%2Fresources%2Fhello.py%3Fcorp%3Dsame-origin due to access control checks.

PASS Same-origin fetch with a 'Cross-Origin-Resource-Policy: same-origin' response header.
PASS Same-origin fetch with a 'Cross-Origin-Resource-Policy: same-site' response header.
PASS Cross-origin cors fetch with a 'Cross-Origin-Resource-Policy: same-origin' response header.
PASS Cross-origin cors fetch with a 'Cross-Origin-Resource-Policy: same-site' response header.
PASS Cross-origin no-cors fetch with a 'Cross-Origin-Resource-Policy: same-origin' response header.
PASS Cross-origin no-cors fetch with a 'Cross-Origin-Resource-Policy: same-site' response header.
PASS Cross-origin no-cors fetch with a 'Cross-Origin-Resource-Policy: same-origin' response header after a redirection.
PASS Cross-origin no-cors fetch with a 'Cross-Origin-Resource-Policy: same-origin' response header after a cross-origin redirection.
PASS Cross-origin no-cors fetch with a 'Cross-Origin-Resource-Policy: same-origin' redirect response header.
